sim: renames around cPar: longValue()->intValue(), cLongParImpl->cIntParImpl, LONG->INT
motivation:
- to be consistent with NED that names integer parameter type "int"
- underlying C++ data type is no longer long but intpar_t (which is int64_t)

+++ src/envir/envirbase.cc
@@ -1549,7 +1549,7 @@ void EnvirBase::setupRNGMapping(cComponent *component)
                 cNEDValue tmp = expr.evaluate(component);
                 if (!tmp.isNumeric())
                     throw opp_runtime_error("Numeric constant or expression expected");
-                physRng = tmp.longValue();
+                physRng = tmp;
             }
 
             if (physRng >= getNumRNGs())
